backend/requirements.txt
# FastAPI & Server
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9

# AI & LLM
openai==1.50.0
langchain==0.3.0
langchain-openai==0.2.0
langchain-pinecone==0.2.0

# Vector Database
pinecone-client==5.0.0

# Data Processing
pydantic==2.9.0
python-dotenv==1.0.1
httpx==0.27.0

# Route Optimization
networkx==3.3
numpy>=1.26.0,<2.0.0

# Utilities
python-dateutil==2.9.0
